Version 0.2.3
Solely changes to back-end.

Simulation class broken into subsequent classes:

Instance:
A final immutable instance comprises of a list of Guests and Rooms. An
inner builder class verifies uniqueness of guest IDs, arrival positions,
and room numbers.

InstanceFactory:
Creates an instance randomly or from a set of CSV files. The static CSV
reader methods are also in this class.

Assignment:
Maintains an assignment of rooms to guests in the form of a HashMap. All
methods used to make a room assignment are in this class. Additionally,
this class maintains the statistics for the current assignment.

Solver:
An interface for a solver which has one method solve; solve takes an
Instance and returns an Assignment.

The previous room assignment methods are now individual classes in the
Solvers package which all implement Solver. Room assignment methods that
only run a single .mod file in AMPL utilize the IPSolver class. An
IPSolver is constructed by specifying a .mod file to run.