Version 0.4.1
Fixed resource leak. In the AMPL API, ampl.finalize() clears some
resources. This is called when the object is deleted but not when the
session is closed. To fix this issue, the API was updated from 2.0.0.3
to 2.0.3.0 and the java garbage collector was called manually.

The parameters for randomly generated instances are now stored in a
.properties file for better readability and efficient modification.

Smaller changes include:
- Guest is now Comparable by guest ID
- Room can only be compared to objects of the same class
- Change in implementation of housekeeper minimum
- Simulation classes renamed
- SimAddGuest output slightly modified
- SimAddGuest progress counter corrected
- Instances now required to have at least one guest and one room
- Updated Tester