Version 0.4.3
NOTE: The GUI is unstable in version 0.4.3

The Instance class is broken into subsequent classes:

Hotel:
A final immutable instance comprises of a list of Rooms and the number
of housekeepers at the hotel. Furthermore, there is a HotelFactory which
can create a Hotel both randomly and from a CSV file.

Instance:
A final immutable instance. The list of Rooms and number of housekeepers
has been removed and replaced by a Hotel. The builder for an instance
takes either a hotel or a copy of an instance. Hence, a hotel must be
built first and an Instance must be built "on" a hotel. The
InstanceFactory can create Instances both randomly and from a CSV file.

Room Qualities are now implemented differently. Now, a Room has an
intrinsic quality (rather than being the average satisfaction of
feasible guests). The CSV writer/reader methods are adjusted
accordingly. This intrinsic quality is used by the BestFirst and
WorstFirst assignment heuristics.

The CSV file structure has changed. Test instances are now in a Test
Files directory. The Simulations directory is no longer located in the
AMPL directory. Using the System.identityHashCode, a unique identifier
is given to each hotel and instance. An instance with id i on a hotel
with id h is represented as follows. A directory named h is located in
the Simulations directory. It contains the CSV file hotel.csv and a
directory called i which contains an arrivals.csv and weights.csv file.

Lastly, the model used to randomly generate hotels and instances has
changed. The randParam.properties file is changed accordingly.
Furthermore, there is now a static class RandProperties which reads and
stores the model parameters upon initialization.